Move RAINBOW_HOME to constants
