+++ src/components/DappBrowser/search/SearchBar.tsx
@@ -15,8 +15,8 @@ import { IS_IOS } from '@/env';
 import { useKeyboardHeight, useDimensions } from '@/hooks';
 import * as i18n from '@/languages';
 import { TAB_BAR_HEIGHT } from '@/navigation/SwipeNavigator';
-import { RAINBOW_HOME, useBrowserContext } from '../BrowserContext';
-import { GOOGLE_SEARCH_URL, HTTP, HTTPS } from '../constants';
+import { useBrowserContext } from '../BrowserContext';
+import { GOOGLE_SEARCH_URL, HTTP, HTTPS, RAINBOW_HOME } from '../constants';
 import { AccountIcon } from '../search-input/AccountIcon';
 import { SearchInput } from '../search-input/SearchInput';
 import { TabButton } from '../search-input/TabButton';
